backend/forecast_service/requirements.txt
SQLAlchemy==1.4.46
psycopg2-binary==2.9.10
python-dotenv==0.21.1
redis==7.0.1
python-dateutil==2.8.2

numpy==2.3.4
pandas==2.3.3
scikit_learn==1.7.2
statsmodels==0.14.5
jsonschema==4.25.1
torch==2.9.1